Start queue implementation and shift to one page design
My idea here is to move to a "One page" design, where there isn't
a profile, household, and home page. Instead, there will be one home
page and the sidebar will contain the users many households.

To run our scheduled reminders, I'm starting with bullMQ. This should
allow us to easily implement a redis-based queue that will run our
scheduled jobs based on a CRON schedule. Eventually this can be used to
send emails or text messages or push notifications. Each time a chore is created,
we should create a job in the queue. Each time a chore is deleted, we should delete.
Edit, edit, etc.